Revert "NumericAddressedBackend: Fix 64->32bit legacy wrapper"
This reverts commit f1122b79cbe5ec71c05647b1e3671ea6a17d228b.

Reason:
- Is already in the master. Will cause unnecessary merge conflicts.

+++ device_backends/src/NumericAddressedBackend.cc
@@ -74,11 +74,11 @@ namespace ChimeraTK {
 
   /* Call 32-bit address implementation by default, for backends that don't implement 64-bit */
   void NumericAddressedBackend::read(uint8_t bar, uint64_t address, int32_t* data, size_t sizeInBytes) {
-    read(bar, static_cast<uint32_t>(address), data, sizeInBytes);
+    read(bar, address, data, sizeInBytes);
   }
 
   void NumericAddressedBackend::write(uint8_t bar, uint64_t address, int32_t const* data, size_t sizeInBytes) {
-    write(bar, static_cast<uint32_t>(address), data, sizeInBytes);
+    write(bar, address, data, sizeInBytes);
   }
 
   /********************************************************************************************************************/
